Changed to float precision in drawmethods Enabled anti-aliasing for some Paint (bombs specifically) Fixed a rounding error for detecting touched row/column

+++ src/com/gmail/wazappdotgithub/ships/model/views/WaitBoardView.java
@@ -25,7 +25,7 @@ public class WaitBoardView extends BoardView {
 	}
 
 	@Override
-	protected void drawSpecial(Canvas canvas, int offset) {
+	protected void drawSpecial(Canvas canvas, float offset) {
 		//Paint my own ships
 		for (IShip s : RemoteClient.getInstance().getBoard().arrayOfShips()) {
 			drawShip(canvas, offset, s);
